+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/configuration/Configuration.java
@@ -71,16 +71,17 @@ public abstract class Configuration {
         }
     }
 
-    public String getUsername() {
-        String username = this.properties.getProperty(USERNAME);
-        return StringUtils.isEmpty(username) ? username : username.replaceAll("@.*", "");
-    }
+	public String getUsername() {
+		String username = this.properties.getProperty(USERNAME);
+		return StringUtils.isEmpty(username) ? username
+			: (username.indexOf("@") > 0 ?  username.substring(0, username.lastIndexOf("@")) : username);
+	}
 
-    public String getDomain() {
-        String username = this.properties.getProperty(USERNAME);
-        return StringUtils.isEmpty(username) ? username
-                : (username.indexOf("@") > 0 ? username.substring(username.indexOf("@") + 1) : null);
-    }
+	public String getDomain() {
+		String username = this.properties.getProperty(USERNAME);
+		return StringUtils.isEmpty(username) ? username
+			: (username.indexOf("@") > 0 ? username.substring(username.lastIndexOf("@") + 1) : null);
+	}
 
     public String getPassword() {
         return this.properties.getProperty(PASSWORD);