+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/configuration/ConfigurationVroTest.java
@@ -18,11 +18,16 @@ package com.vmware.pscoe.iac.artifact.configuration;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Properties;
+import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
+import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 
 class ConfigurationVroTest {
@@ -48,6 +53,44 @@ class ConfigurationVroTest {
 		this.properties.remove("username");
 		this.properties.remove("password");
 		this.properties.remove(AUTH_METHOD);
+	}
+
+	@ParameterizedTest
+	@MethodSource("usernameProvider")
+	public void testGetUsername(String username, String auth, String expectedUsername, String expectedDomain ) {
+		Properties properties = new Properties();
+		properties.setProperty("username", username);
+		properties.setProperty("auth", auth);
+		ConfigurationVro configuration = new ConfigurationVro(properties);
+
+		assertEquals(expectedUsername, configuration.getUsername());
+	}
+	@ParameterizedTest
+	@MethodSource("usernameProvider")
+	public void testGetDomain(String username, String auth, String expectedUsername, String expectedDomain ) {
+		Properties properties = new Properties();
+		properties.setProperty("username", username);
+		properties.setProperty("auth", auth);
+		ConfigurationVro configuration = new ConfigurationVro(properties);
+
+		assertEquals(expectedDomain, configuration.getDomain());
+	}
+
+	private static Stream<Arguments> usernameProvider() {
+		return Stream.of(
+			arguments("configurationadmin", "vra", "configurationadmin", null),
+			arguments("configurationadmin@corp.local", "vra", "configurationadmin", "corp.local"),
+			arguments("configurationadmin@System Domain", "vra", "configurationadmin", "System Domain"),
+			arguments("configurationadmin@@System Domain", "vra", "configurationadmin@", "System Domain"),
+			arguments("configurationadmin@test@System Domain", "vra", "configurationadmin@test", "System Domain"),
+			arguments("configurationadmin@test@test2@System Domain", "vra", "configurationadmin@test@test2", "System Domain"),
+			arguments("configurationadmin", "basic", "configurationadmin", null),
+			arguments("configurationadmin@corp.local", "basic", "configurationadmin@corp.local", "corp.local"),
+			arguments("configurationadmin@System Domain", "basic", "configurationadmin@System Domain", "System Domain"),
+			arguments("configurationadmin@@System Domain", "basic", "configurationadmin@@System Domain", "System Domain"),
+			arguments("configurationadmin@test@System Domain", "basic", "configurationadmin@test@System Domain", "System Domain"),
+			arguments("configurationadmin@test@test2@System Domain", "basic", "configurationadmin@test@test2@System Domain", "System Domain")
+		);
 	}
 
 	@Test